modify login when add to cart

+++ src/app/Service/product.service.ts
@@ -57,22 +57,32 @@ export class ProductService {
 
 
   public addProductToWishList(product: Product): void {
+    if(!this.authService.IsAuthenticated()){
+        return ;
+    }   
     this.http.post(`${this.apiUrl}/Baskets/favourite`, product).subscribe();
   }
 
 
   public removeProductFromWishList(productId: number): void {
+    if(!this.authService.IsAuthenticated()){
+        return ;
+    }
     this.http.delete(`${this.apiUrl}/Baskets/favourite/${productId}`).subscribe();
   }
 
 
   public getProductByID(id: number): Observable<Product> {
+    if(!this.authService.IsAuthenticated()){
+        return EMPTY;
+    }
     return this.http.get<Product>(`${this.apiUrl}/Products/${id}`);
   }
 
   public addToCart(product: Product, quantity: number): Observable<any> {
 
     if(!this.authService.IsAuthenticated()){
+      console.log("not authorize");
       return EMPTY;
     }
     const payload = {
